+++ app/build.gradle
@@ -24,6 +24,9 @@ dependencies {
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    compile project(':activation')
+    compile project(':additionnal')
+    compile project(':mail')
     compile 'com.android.support:appcompat-v7:26.0.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.google.firebase:firebase-auth:11.2.0'
@@ -33,10 +36,8 @@ dependencies {
     compile 'com.google.firebase:firebase-database:11.2.0'
     compile 'com.android.support:cardview-v7:26.0.1'
     compile 'com.android.support:recyclerview-v7:26.0.1'
-    testCompile 'junit:junit:4.12'
-    compile project(':activation')
-    compile project(':additionnal')
-    compile project(':mail')
     compile 'com.mobsandgeeks:android-saripaar:2.0.3'
+    compile 'com.android.support:design:26.0.1'
+    testCompile 'junit:junit:4.12'
 }
 apply plugin: 'com.google.gms.google-services'
