+++ app/src/main/java/com/example/shenawynkov/jopfinder/LoginActivity.java
@@ -4,6 +4,7 @@ import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -50,7 +51,8 @@ public class LoginActivity extends BaseActivity implements Validator.ValidationL
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
-
+        Toolbar myToolbar = (Toolbar) findViewById(R.id.my_toolbar);
+        setSupportActionBar(myToolbar);
         mValidator = new Validator(this);
         mValidator.setValidationListener(this);
 
@@ -145,7 +147,7 @@ public class LoginActivity extends BaseActivity implements Validator.ValidationL
 
                             startActivity(intent);
                         } else {
-                            Intent intent = new Intent(LoginActivity.this, NavigationActivity.class);
+                            Intent intent = new Intent(LoginActivity.this, JopListActivity.class);
                             intent.putExtra(getString(R.string.user_extra),user1);
                             startActivity(intent);
