+++ app/src/main/res/layout/activity_new_jop.xml
@@ -1,12 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.design.widget.CoordinatorLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.example.shenawynkov.jopfinder.NewJopActivity"
-    android:orientation="vertical">
+>
+    <android.support.design.widget.AppBarLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay">
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/my_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:elevation="4dp"
+        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
+    </android.support.design.widget.AppBarLayout>
+    <LinearLayout
+        android:id="@+id/linear"
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="?actionBarSize">
     <EditText
         android:id="@+id/newjop_title"
         android:layout_width="match_parent"
@@ -74,5 +93,17 @@ android:layout_margin="16dp"
         android:layout_marginBottom="16dp"
         android:layout_gravity="center_horizontal"
         android:layout_margin="16dp"/>
+    </LinearLayout>
 
-</LinearLayout>
+<android.support.design.widget.FloatingActionButton
+android:layout_width="wrap_content"
+android:layout_height="wrap_content"
+android:layout_gravity="bottom|right"
+android:layout_margin="16dp"
+android:elevation="6dp"
+app:pressedTranslationZ="12dp"
+android:src="@drawable/add"
+app:layout_anchor="@id/linear"
+app:layout_anchorGravity="bottom|right|end" />
+</android.support.design.widget.CoordinatorLayout>
+
