+++ app/src/main/res/layout/activity_sign_up.xml
@@ -1,11 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
+<android.support.design.widget.CoordinatorLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <android.support.design.widget.AppBarLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay">
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/my_toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"
+            android:elevation="4dp"
+            android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+            app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
+    </android.support.design.widget.AppBarLayout>
+
+
+<android.support.constraint.ConstraintLayout
+    android:layout_marginTop="?actionBarSize"
+    android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.example.shenawynkov.jopfinder.SignUpActivity">
+
 
     <EditText
         android:id="@+id/email_sign_up"
@@ -102,3 +122,4 @@
         android:layout_marginTop="14dp"
         app:layout_constraintTop_toBottomOf="@+id/spinner_signup"/>
 </android.support.constraint.ConstraintLayout>
+</android.support.design.widget.CoordinatorLayout>
