--- app/src/main/res/layout/fragment_selection.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-  >
-
-    <Button
-        android:id="@+id/employer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="employer"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="68dp"
-        android:layout_marginRight="148dp"
-        app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginLeft="8dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"/>
-
-    <Button
-        android:id="@+id/JobSeeker"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="JobSeeker"
-        android:layout_marginTop="77dp"
-        app:layout_constraintTop_toBottomOf="@+id/employer"
-        android:layout_marginRight="148dp"
-        app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginLeft="8dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"/>
-</android.support.constraint.ConstraintLayout>
