+++ app/src/main/res/menu/new_jop_menu.xml
@@ -2,8 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        android:title="@string/action_settings"
-        app:showAsAction="never"/>
+        android:id="@+id/action_logout"
+        android:title="@string/sign_out"
+        app:showAsAction="ifRoom"/>
 </menu>
