+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,6 @@
 <resources>
     <string name="app_name">Job Finder</string>
-    <string name="title_activity_navigation">NavigationActivity</string>
 
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
     <string name="user_extra"> user extra</string>
